+++ ntex/src/client/response.rs
@@ -1,5 +1,5 @@
 use std::cell::{Cell, Ref, RefMut};
-use std::task::{Context, Poll};
+use std::task::{Context, Poll, ready};
 use std::{fmt, future::Future, marker::PhantomData, pin::Pin};
 
 use serde::de::DeserializeOwned;
@@ -7,13 +7,15 @@ use serde::de::DeserializeOwned;
 #[cfg(feature = "cookie")]
 use coo_kie::{Cookie, ParseError as CookieParseError};
 
+use crate::error::Error;
 use crate::http::error::PayloadError;
 use crate::http::header::{AsName, CONTENT_LENGTH, HeaderValue};
 use crate::http::{HeaderMap, HttpMessage, Payload, ResponseHead, StatusCode, Version};
 use crate::time::{Deadline, Millis};
 use crate::util::{Bytes, BytesMut, Extensions, Stream};
 
-use super::{ClientConfig, ServiceResponse, error::JsonPayloadError};
+use super::error::{ClientPayloadError, JsonPayloadError};
+use super::{ClientConfig, ServiceResponse};
 
 /// Client Response
 pub struct ClientResponse {
@@ -160,13 +162,16 @@ impl ClientResponse {
 }
 
 impl Stream for ClientResponse {
-    type Item = Result<Bytes, PayloadError>;
+    type Item = Result<Bytes, Error<ClientPayloadError>>;
 
     fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         if let Some(mut pl) = self.payload.take() {
             let result = Pin::new(&mut pl).poll_next(cx);
             self.payload.set(Some(pl));
-            result
+            Poll::Ready(
+                ready!(result)
+                    .map(|item| item.map_err(|e| Error::from(ClientPayloadError(e)))),
+            )
         } else {
             Poll::Ready(None)
         }
@@ -198,23 +203,34 @@ impl fmt::Debug for ClientResponse {
 /// Future that resolves to a complete http message body.
 pub struct MessageBody {
     length: Option<usize>,
-    err: Option<PayloadError>,
+    err: Option<Error<ClientPayloadError>>,
     fut: Option<ReadBody>,
+    config: ClientConfig,
 }
 
 impl MessageBody {
     /// Create `MessageBody` for request.
     pub fn new(res: &ClientResponse) -> MessageBody {
+        let config = res.config.clone();
+
         let mut len = None;
         if let Some(l) = res.headers().get(&CONTENT_LENGTH) {
             if let Ok(s) = l.to_str() {
                 if let Ok(l) = s.parse::<usize>() {
                     len = Some(l);
                 } else {
-                    return Self::err(PayloadError::UnknownLength);
+                    return Self::err(
+                        Error::from(ClientPayloadError(PayloadError::UnknownLength))
+                            .set_service(config.cfg().service()),
+                        config,
+                    );
                 }
             } else {
-                return Self::err(PayloadError::UnknownLength);
+                return Self::err(
+                    Error::from(ClientPayloadError(PayloadError::UnknownLength))
+                        .set_service(config.cfg().service()),
+                    config,
+                );
             }
         }
 
@@ -225,7 +241,9 @@ impl MessageBody {
                 res.take_payload(),
                 res.config.payload_limit(),
                 res.config.payload_timeout(),
+                res.config.clone(),
             )),
+            config: res.config.clone(),
         }
     }
 
@@ -252,8 +270,9 @@ impl MessageBody {
         self
     }
 
-    fn err(e: PayloadError) -> Self {
+    fn err(e: Error<ClientPayloadError>, config: ClientConfig) -> Self {
         MessageBody {
+            config,
             fut: None,
             err: Some(e),
             length: None,
@@ -262,7 +281,7 @@ impl MessageBody {
 }
 
 impl Future for MessageBody {
-    type Output = Result<Bytes, PayloadError>;
+    type Output = Result<Bytes, Error<ClientPayloadError>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let this = self.get_mut();
@@ -274,11 +293,16 @@ impl Future for MessageBody {
         if let Some(len) = this.length.take() {
             let limit = this.fut.as_ref().unwrap().limit;
             if limit > 0 && len > limit {
-                return Poll::Ready(Err(PayloadError::Overflow));
+                return Poll::Ready(Err(Error::from(ClientPayloadError(
+                    PayloadError::Overflow,
+                ))
+                .set_service(this.config.cfg().service())));
             }
         }
 
-        Pin::new(&mut this.fut.as_mut().unwrap()).poll(cx)
+        Pin::new(&mut this.fut.as_mut().unwrap())
+            .poll(cx)
+            .map_err(|e| e.map(ClientPayloadError::from))
     }
 }
 
@@ -291,8 +315,9 @@ impl Future for MessageBody {
 /// * content length is greater than 64k
 pub struct JsonBody<U> {
     length: Option<usize>,
-    err: Option<JsonPayloadError>,
+    err: Option<Error<JsonPayloadError>>,
     fut: Option<ReadBody>,
+    config: ClientConfig,
     _t: PhantomData<U>,
 }
 
@@ -303,6 +328,8 @@ where
     #[must_use]
     /// Create `JsonBody` for request.
     pub fn new(res: &ClientResponse) -> Self {
+        let config = res.config.clone();
+
         // check content-type
         let json = if let Ok(Some(mime)) = res.mime_type() {
             mime.subtype() == mime::JSON || mime.suffix() == Some(mime::JSON)
@@ -310,10 +337,16 @@ where
             false
         };
         if !json {
+            let err = Some(
+                Error::from(JsonPayloadError::ContentType)
+                    .set_service(config.cfg().service()),
+            );
+
             return JsonBody {
+                err,
+                config,
                 length: None,
                 fut: None,
-                err: Some(JsonPayloadError::ContentType),
                 _t: PhantomData,
             };
         }
@@ -327,12 +360,14 @@ where
         }
 
         JsonBody {
+            config,
             length: len,
             err: None,
             fut: Some(ReadBody::new(
                 res.take_payload(),
                 res.config.payload_limit(),
                 res.config.payload_timeout(),
+                res.config.clone(),
             )),
             _t: PhantomData,
         }
@@ -368,7 +403,7 @@ impl<U> Future for JsonBody<U>
 where
     U: DeserializeOwned,
 {
-    type Output = Result<U, JsonPayloadError>;
+    type Output = Result<U, Error<JsonPayloadError>>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         if let Some(err) = self.err.take() {
@@ -378,17 +413,21 @@ where
         if let Some(len) = self.length.take() {
             let limit = self.fut.as_ref().unwrap().limit;
             if limit > 0 && len > limit {
-                return Poll::Ready(Err(JsonPayloadError::Payload(
-                    PayloadError::Overflow.into(),
-                )));
+                return Poll::Ready(Err(Error::from(JsonPayloadError::Payload(
+                    ClientPayloadError(PayloadError::Overflow),
+                ))
+                .set_service(self.config.cfg().service())));
             }
         }
 
-        let body = match Pin::new(&mut self.get_mut().fut.as_mut().unwrap()).poll(cx) {
-            Poll::Ready(result) => result?,
+        let this = self.get_mut();
+        let body = match Pin::new(&mut this.fut.as_mut().unwrap()).poll(cx) {
+            Poll::Ready(result) => result.map_err(|e| e.map(JsonPayloadError::from))?,
             Poll::Pending => return Poll::Pending,
         };
-        Poll::Ready(serde_json::from_slice::<U>(&body).map_err(JsonPayloadError::from))
+        Poll::Ready(serde_json::from_slice::<U>(&body).map_err(|e| {
+            Error::from(JsonPayloadError::from(e)).set_service(this.config.cfg().service())
+        }))
     }
 }
 
@@ -398,13 +437,15 @@ struct ReadBody {
     buf: BytesMut,
     limit: usize,
     timeout: Deadline,
+    config: ClientConfig,
 }
 
 impl ReadBody {
-    fn new(stream: Payload, limit: usize, timeout: Millis) -> Self {
+    fn new(stream: Payload, limit: usize, timeout: Millis, config: ClientConfig) -> Self {
         Self {
             stream,
             limit,
+            config,
             buf: BytesMut::with_capacity(std::cmp::min(limit, 32768)),
             timeout: Deadline::new(timeout),
         }
@@ -412,7 +453,7 @@ impl ReadBody {
 }
 
 impl Future for ReadBody {
-    type Output = Result<Bytes, PayloadError>;
+    type Output = Result<Bytes, Error<ClientPayloadError>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let this = self.get_mut();
@@ -421,22 +462,29 @@ impl Future for ReadBody {
             return match Pin::new(&mut this.stream).poll_next(cx) {
                 Poll::Ready(Some(Ok(chunk))) => {
                     if this.limit > 0 && (this.buf.len() + chunk.len()) > this.limit {
-                        Poll::Ready(Err(PayloadError::Overflow))
+                        Poll::Ready(Err(Error::from(ClientPayloadError(
+                            PayloadError::Overflow,
+                        ))
+                        .set_service(this.config.cfg().service())))
                     } else {
                         this.buf.extend_from_slice(&chunk);
                         continue;
                     }
                 }
                 Poll::Ready(None) => Poll::Ready(Ok(this.buf.take())),
-                Poll::Ready(Some(Err(err))) => Poll::Ready(Err(err)),
+                Poll::Ready(Some(Err(err))) => {
+                    Poll::Ready(Err(Error::from(ClientPayloadError(err))
+                        .set_service(this.config.cfg().service())))
+                }
                 Poll::Pending => {
                     if this.timeout.poll_elapsed(cx).is_ready() {
-                        Poll::Ready(Err(PayloadError::Incomplete(Some(
-                            std::io::Error::new(
+                        Poll::Ready(Err(Error::from(ClientPayloadError(
+                            PayloadError::Incomplete(Some(std::io::Error::new(
                                 std::io::ErrorKind::TimedOut,
                                 "Operation timed out",
-                            ),
-                        ))))
+                            ))),
+                        ))
+                        .set_service(this.config.cfg().service())))
                     } else {
                         Poll::Pending
                     }
@@ -457,13 +505,13 @@ mod tests {
     #[crate::rt_test]
     async fn test_body() {
         let req = TestResponse::with_header(header::CONTENT_LENGTH, "xxxx").finish();
-        match req.body().await.err().unwrap() {
+        match &*req.body().await.err().unwrap().into_error() {
             PayloadError::UnknownLength => (),
             _ => unreachable!("error"),
         }
 
         let req = TestResponse::with_header(header::CONTENT_LENGTH, "1000000").finish();
-        match req.body().await.err().unwrap() {
+        match &*req.body().await.err().unwrap().into_error() {
             PayloadError::Overflow => (),
             _ => unreachable!("error"),
         }
@@ -476,7 +524,7 @@ mod tests {
         let req = TestResponse::default()
             .set_payload(Bytes::from_static(b"11111111111111"))
             .finish();
-        match req.body().limit(5).await.err().unwrap() {
+        match &*req.body().limit(5).await.err().unwrap().into_error() {
             PayloadError::Overflow => (),
             _ => unreachable!("error"),
         }