+++ ntex/tests/http_awc_client.rs
@@ -10,9 +10,10 @@ use ntex::http::test::server as test_server;
 use ntex::http::{HttpMessage, HttpService, header};
 use ntex::io::IoConfig;
 use ntex::service::{cfg::SharedCfg, chain_factory, fn_layer};
+use ntex::time::{Millis, Seconds, sleep};
 use ntex::web::middleware::Compress;
 use ntex::web::{self, App, BodyEncoding, Error, HttpRequest, HttpResponse, test};
-use ntex::{client, time::Millis, time::Seconds, time::sleep, util::Bytes, util::Ready};
+use ntex::{client, error, util::Bytes, util::Ready};
 
 const STR: &str = "Hello World Hello World Hello World Hello World Hello World \
                    Hello World Hello World Hello World Hello World Hello World \
@@ -106,7 +107,7 @@ async fn test_timeout() {
     })
     .await;
 
-    let connector = Connector::default().connector(ntex::connect::Connector::new());
+    let connector = Connector::default().connector(ntex::connect::Connector2::new());
 
     let client = Client::builder()
         .connector::<&str>(connector)
@@ -116,7 +117,7 @@ async fn test_timeout() {
         .unwrap();
 
     let request = client.get(srv.url("/")).send().await;
-    match request {
+    match request.map_err(error::Error::into_error) {
         Err(ClientError::Timeout) => (),
         _ => panic!(),
     }
@@ -138,7 +139,7 @@ async fn test_timeout_override() {
         .await
         .unwrap();
     let request = client.get(srv.url("/")).timeout(Seconds(1)).send();
-    match request.await {
+    match request.await.map_err(error::Error::into_error) {
         Err(ClientError::Timeout) => (),
         _ => panic!(),
     }